Small final pre-style refinements -- TOC + splash header links, toc item links
