+++ packages/11ty/_includes/components/download-link.js
@@ -3,6 +3,9 @@ const path = require('path')
 const { oneLine } = require('~lib/common-tags')
 const chalkFactory = require('~lib/chalk')
 
+const checkFormat = require('../../_plugins/collections/filters/output.js')
+
+const logger = chalkFactory("shortcode:footer-dl")
 /**
  * Handles logic about whether a footer d/l link appears at all and renders a link
  *
@@ -21,14 +24,45 @@ module.exports = function(eleventyConfig) {
 
   const pdfConfig = eleventyConfig.globalData.config.pdf
 
-  return function (params) {
-    const { key, label } = params
+  /**
+   * @function checkLinkPagePDF
+   * 
+   * @param {Object} config pdf object from Quire config
+   * @param {Array<string>,string,undefined} outputs outputs setting from page frontmatter 
+   * @param {bool} frontmatterSetting pdf page setting from page frontmatter
+   * 
+   * Returns true if a download link of this kind is configured
+   **/
+  const checkLinkPagePDF = (type,config,outputs,frontmatterSetting) => {
 
+    // Is the output being created?
+    if ( !checkFormat('pdf',{data:{outputs}}) ) { 
+      return false 
+    }
+
+    // Are the links of this type set?
+    if ( config.pagePDF.accessLinks.find( (al) => al[type] === true ) === undefined )  {
+      return false
+    }
+
+    // Return the core logic check
+    return ( config.pagePDF.output === true && frontmatterSetting !== false ) || frontmatterSetting === true
+
+  }
+
+  return function (params) {
+    const { key, outputs, page_pdf_output: pagePDFOutput, type } = params
+
+    if (!checkLinkPagePDF(type,pdfConfig,outputs,pagePDFOutput)) { 
+      return ''
+    }
+
+    const text = pdfConfig.pagePDF.accessLinks.find( al => al[type] === true ).label
     const href = path.join( pdfConfig.outputDir, `${pdfConfig.filename}-${slugify(key)}.pdf` )
 
     return oneLine`
       <div class="quire-download" data-outputs-exclude="epub,pdf">
-        <a class="quire-download__link" href="${ href }" target="_blank" ><span>${ label }</span><svg class="quire-download__link__icon"><use xlink:href="#download-icon"></use></svg></a>
+        <a class="quire-download__link" href="${ href }" target="_blank" ><span>${ text }</span><svg class="quire-download__link__icon"><use xlink:href="#download-icon"></use></svg></a>
       </div>`
   }
