+++ packages/11ty/_includes/components/index.js
@@ -33,7 +33,6 @@ module.exports = {
   figureTableElement: require('./figure/table/element'),
   figureVideo: require('./figure/video'),
   figureVideoElement: require('./figure/video/element'),
-  footerDownloadLink: require('./footer-download-link.js'),
   head: require('./head.js'),
   icon: require('./icon.js'),
   icons: require('./icons.js'),
