+++ packages/11ty/_layouts/entry.liquid
@@ -68,7 +68,8 @@ Entry content, including entry image and tombstone data
         <div class="content">
           {{ content }}
           {% bibliography citations %}
-          {% footerDownloadLink 
+          {% downloadLink 
+             type="footer"
              key=key
              outputs=outputs 
              page_pdf_output=page_pdf_output
