+++ packages/11ty/_layouts/splash.liquid
@@ -24,9 +24,6 @@ description: splash page layout
         {% contributors context=pageContributors, format=byline_format %}
       </div>
     {% endif %}
-    {% comment %}
-    FIXME: add header download link
-    {% endcomment %}
   </div>
 </section>
 
@@ -37,12 +34,19 @@ description: splash page layout
     {% endif %}
     <div class="container">
       <div class="content{% if image %}{% else %} no-image-above{% endif %}">
+        {% downloadLink
+          type="header"
+          key=key
+          outputs=outputs
+          page_pdf_output=page_pdf_output
+        %}
         {{ content }}
         {% bibliography citations %}
-        {% footerDownloadLink 
-         key=key
-         outputs=outputs 
-         page_pdf_output=page_pdf_output
+        {% downloadLink 
+          type="footer"
+          key=key
+          outputs=outputs 
+          page_pdf_output=page_pdf_output
         %}
       </div>
       {% pageButtons pagination=pagination %}
