+++ packages/11ty/content/_computed/eleventyComputed.js
@@ -27,6 +27,7 @@ module.exports = {
         layout: data.layout,
         object: data.object,
         order: data.order,
+        page_pdf_output: data.page_pdf_output,
         short_title: data.short_title,
         subtitle: data.subtitle,
         summary: data.summary,
